+++ includes/cpp_redis/network/redis_connection.hpp
@@ -6,10 +6,10 @@
 #include <functional>
 
 #ifdef _MSC_VER
-#include <cpp_redis/network/win_tcp_client.hpp>
+# include <cpp_redis/network/windows/tcp_client.hpp>
 #else
-#include <cpp_redis/network/tcp_client.hpp>
-#endif
+# include <cpp_redis/network/unix/tcp_client.hpp>
+#endif /* _MSC_VER */
 
 #include <cpp_redis/builders/reply_builder.hpp>
 
@@ -20,7 +20,7 @@ namespace network {
 class redis_connection {
 public:
   //! ctor & dtor
-  redis_connection::redis_connection(const std::shared_ptr<io_service> pIO);
+  redis_connection(const std::shared_ptr<io_service>& IO);
   ~redis_connection(void);
 
   //! copy ctor & assignment operator
