+++ includes/cpp_redis/redis_client.hpp
@@ -15,7 +15,7 @@ namespace cpp_redis {
 class redis_client {
 public:
   //! ctor & dtor
-  redis_client(const std::shared_ptr<network::io_service> pIO =NULL);
+  redis_client(const std::shared_ptr<network::io_service>& IO = nullptr);
   ~redis_client(void);
 
   //! copy ctor & assignment operator
