+++ sources/network/redis_connection.cpp
@@ -5,10 +5,10 @@ namespace cpp_redis {
 
 namespace network {
 
-redis_connection::redis_connection(const std::shared_ptr<io_service> pIO/*=NULL*/)
-: m_reply_callback(nullptr)
+redis_connection::redis_connection(const std::shared_ptr<io_service>& IO)
+: m_client(IO)
+, m_reply_callback(nullptr)
 , m_disconnection_handler(nullptr)
-, m_client(pIO)
 {
   __CPP_REDIS_LOG(debug, "cpp_redis::network::redis_connection created");
 }
